backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.26.5
django-filter==23.5

django-storages[boto3]==1.14.2
boto3==1.34.34
django-cloudinary-storage==0.3.0
cloudinary==1.41.0